Move fallback class
