+++ src/main/java/org/opentripplanner/api/common/RoutingResource.java
@@ -19,9 +19,9 @@ import javax.xml.datatype.DatatypeConstants;
 import javax.xml.datatype.DatatypeFactory;
 import javax.xml.datatype.XMLGregorianCalendar;
 import org.opentripplanner.api.parameter.QualifiedModeSet;
+import org.opentripplanner.api.support.ZoneIdFallback;
 import org.opentripplanner.ext.dataoverlay.api.DataOverlayParameters;
 import org.opentripplanner.framework.application.OTPFeature;
-import org.opentripplanner.framework.application.ZoneIdFallback;
 import org.opentripplanner.framework.lang.StringUtils;
 import org.opentripplanner.framework.time.DurationUtils;
 import org.opentripplanner.routing.api.request.RouteRequest;
